ruangpeduli/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
cryptography==46.0.5
dj-database-url==3.1.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.0
djangorestframework-simplejwt==5.4.0
google-auth==2.49.1
idna==3.11
pillow==12.1.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
python-dotenv==1.2.2
requests==2.32.5
resend==2.26.0
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
